feat: add public wallet and fix stringify tx
