+++ packages/contract/src/contracts/functions/base-invocation-scope.ts
@@ -2,7 +2,7 @@
 import type { InputValue } from '@fuel-ts/abi-coder';
 import type { ContractIdLike } from '@fuel-ts/interfaces';
 import { bn, toNumber } from '@fuel-ts/math';
-import type { Provider, CoinQuantity } from '@fuel-ts/providers';
+import type { Provider, CoinQuantity, TransactionRequest } from '@fuel-ts/providers';
 import { transactionRequestify, ScriptTransactionRequest } from '@fuel-ts/providers';
 import { MAX_GAS_PER_TX, InputType } from '@fuel-ts/transactions';
 
@@ -185,6 +185,18 @@ export class BaseInvocationScope<TReturn = any> {
   }
 
   /**
+   * Prepare transaction request object, adding Inputs, Outputs, coins, check gas costs
+   * and transaction validity.
+   *
+   * It's possible to get the transaction without adding coins, by passing `fundTransaction`
+   * as false.
+   */
+  async getTransactionRequest(options?: CallOptions): Promise<TransactionRequest> {
+    await this.prepareTransaction(options);
+    return this.transactionRequest;
+  }
+
+  /**
    * Submits a transaction to the blockchain.
    *
    * This is a final action and will spend the coins and change the state of the contract.
@@ -194,8 +206,8 @@ export class BaseInvocationScope<TReturn = any> {
   async call<T = TReturn>(options?: CallOptions): Promise<FunctionInvocationResult<T>> {
     assert(this.contract.wallet, 'Wallet is required!');
 
-    await this.prepareTransaction(options);
-    const response = await this.contract.wallet.sendTransaction(this.transactionRequest);
+    const transactionRequest = await this.getTransactionRequest(options);
+    const response = await this.contract.wallet.sendTransaction(transactionRequest);
 
     return FunctionInvocationResult.build<T>(
       this.functionInvocationScopes,
@@ -215,8 +227,8 @@ export class BaseInvocationScope<TReturn = any> {
   async simulate<T = TReturn>(options?: CallOptions): Promise<InvocationCallResult<T>> {
     assert(this.contract.wallet, 'Wallet is required!');
 
-    await this.prepareTransaction(options);
-    const result = await this.contract.wallet.simulateTransaction(this.transactionRequest);
+    const transactionRequest = await this.getTransactionRequest(options);
+    const result = await this.contract.wallet.simulateTransaction(transactionRequest);
 
     return InvocationCallResult.build<T>(this.functionInvocationScopes, result, this.isMultiCall);
   }
@@ -233,8 +245,8 @@ export class BaseInvocationScope<TReturn = any> {
     const provider = (this.contract.wallet?.provider || this.contract.provider) as Provider;
     assert(provider, 'Wallet or Provider is required!');
 
-    await this.prepareTransaction(options);
-    const request = transactionRequestify(this.transactionRequest);
+    const transactionRequest = await this.getTransactionRequest(options);
+    const request = transactionRequestify(transactionRequest);
     const response = await provider.call(request, {
       utxoValidation: false,
     });
