+++ packages/fuel-gauge/src/token-test-contract.test.ts
@@ -22,7 +22,7 @@ const setup = async () => {
 };
 
 describe('TokenTestContract', () => {
-  it.only('Can mint and transfer coins', async () => {
+  it('Can mint and transfer coins', async () => {
     // New wallet to transfer coins and check balance
     const userWallet = Wallet.generate({ provider });
     const token = await setup();
