+++ packages/typechain-target-fuels/example/types/factories/Demo__factory.ts
@@ -2,7 +2,7 @@
 /* tslint:disable */
 /* eslint-disable */
 
-import type { Provider, Wallet, AbstractAddress } from "fuels";
+import type { Provider, WalletPublic, AbstractAddress } from "fuels";
 import { Interface, Contract } from "fuels";
 import type { Demo, DemoInterface } from "../Demo";
 const _abi = [
@@ -61,7 +61,7 @@ export class Demo__factory {
   }
   static connect(
     id: string | AbstractAddress,
-    walletOrProvider: Wallet | Provider
+    walletOrProvider: WalletPublic | Provider
   ): Demo {
     return new Contract(id, _abi, walletOrProvider) as unknown as Demo;
   }
