+++ packages/typechain-target-fuels/example/types/factories/Token__factory.ts
@@ -2,7 +2,7 @@
 /* tslint:disable */
 /* eslint-disable */
 
-import type { Provider, Wallet, AbstractAddress } from "fuels";
+import type { Provider, WalletPublic, AbstractAddress } from "fuels";
 import { Interface, Contract } from "fuels";
 import type { Token, TokenInterface } from "../Token";
 const _abi = [
@@ -158,7 +158,7 @@ export class Token__factory {
   }
   static connect(
     id: string | AbstractAddress,
-    walletOrProvider: Wallet | Provider
+    walletOrProvider: WalletPublic | Provider
   ): Token {
     return new Contract(id, _abi, walletOrProvider) as unknown as Token;
   }
