+++ packages/wallet/src/index.ts
@@ -1,2 +1,3 @@
 export { default as Wallet } from './wallet';
+export { default as WalletPublic } from './wallet-public';
 export * as TestUtils from './test-utils';
